+++ .github/workflows/blank.yml
@@ -6,7 +6,10 @@ on:
   pull_request:
     branches: [ "master", "develop" ]
   workflow_dispatch:
-  
+
+env:
+  SELENIUM: on
+
 jobs:
   tests:
     runs-on: ubuntu-latest
