+++ package.json
@@ -3,10 +3,11 @@
   "type": "module",
   "devDependencies": {
     "@wdio/allure-reporter": "^8.5.0",
-    "@wdio/local-runner": "^8.6.8",
     "@wdio/cli": "^8.5.5",
     "@wdio/globals": "^8.5.9",
+    "@wdio/local-runner": "^8.6.8",
     "@wdio/mocha-framework": "^8.4.0",
+    "@wdio/selenium-standalone-service": "^8.6.8",
     "@wdio/types": "^8.4.0",
     "allure-commandline": "^2.21.0",
     "chromedriver": "^111.0.0",
